requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-strip-extras --output-file=requirements.txt requirements.in
#
asgiref==3.11.1
    # via django
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via rq
croniter==6.2.2
    # via rq
cryptography==43.0.3
    # via -r requirements.in
django==5.2.12
    # via
    #   -r requirements.in
    #   django-filter
    #   django-rq
    #   django-tables2
    #   djangorestframework
    #   drf-spectacular
django-filter==25.2
    # via -r requirements.in
django-rq==3.2.2
    # via -r requirements.in
django-tables2==2.8.0
    # via -r requirements.in
djangorestframework==3.17.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.29.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
psycopg==3.3.3
    # via -r requirements.in
pycparser==2.23
    # via cffi
pyjwt==2.12.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via croniter
pyyaml==6.0.3
    # via drf-spectacular
redis==7.4.0
    # via
    #   django-rq
    #   rq
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rq==2.7.0
    # via
    #   -r requirements.in
    #   django-rq
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   psycopg
    #   referencing
uritemplate==4.2.0
    # via drf-spectacular
